refactor: 디렉토리 구조 개편 (서비스 단위 모듈화)
- src/auth/, src/core/ 제거
- src/kis/client.py: KISAuth + KISAgent를 KISClient 단일 클래스로 통합
- src/kis/stock_config.py: KIS 거래소 상수 이동
- src/sheets/client.py: GCPAuth → GoogleSheetsClient로 이동 및 rename
- src/allocation.py: static_allocation.py 이동
- src/logger.py: log_method_call 데코레이터 통합 (config/helper.py 제거)
- main.py: 변경된 import 경로 반영
