--- src/auth/__init__.py
@@ -1,2 +1,0 @@
-from src.auth.gcp_auth import *
-from src.auth.kis_auth import *
