--- src/core/__init__.py
@@ -1,2 +1,0 @@
-from src.core.kis_agent import *
-from src.core.stock_config import *
