+++ src/kis/stock_config.py
